remove leaf parameters, greedy search, and move minimax to separate class
